requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
mysqlclient==2.2.4
pymysql==1.1.0
email-validator==2.1.0
python-dotenv==1.0.1
pytest==8.0.1
httpx==0.26.0
pandas==2.2.0 